requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
cloud-sql-python-connector==0.4.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==35.0.0
Flask==2.0.1
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.28.0
google-cloud-core==2.0.0
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
proto-plus==1.19.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pyarrow==5.0.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
# pypiwin32==223
python-dateutil==2.8.2
pytz==2021.1
# pywin32==301
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.16.0
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
webdriver-manager==3.4.2
Werkzeug==2.0.1
yarl==1.6.3
